Update to current

+++ src/map/item_container.cpp
@@ -146,7 +146,7 @@ uint8 CItemContainer::InsertItem(CItem* PItem)
             m_count++;
 
 			PItem->setSlotID(SlotID);
-			PItem->setLocationID(m_id);
+			PItem->setLocationID((uint8)m_id);
 
 			m_ItemList[SlotID] = PItem;
 			return SlotID;
@@ -171,7 +171,7 @@ uint8 CItemContainer::InsertItem(CItem* PItem, uint8 SlotID)
 		if (PItem != nullptr)
 		{
 			PItem->setSlotID(SlotID);
-			PItem->setLocationID(m_id);
+			PItem->setLocationID((uint8)m_id);
 
             if (m_ItemList[SlotID] == nullptr && SlotID != 0) m_count++;
 		}
